+++ Diary/ForthItem/Controller/MyMeansViewController.m
@@ -75,7 +75,7 @@ typedef NS_ENUM(NSInteger, PhotoType)
         
         //打印数据
         
-      NSLog(@"ID : %ld, name : %@, age : %@ ",model.id,model.name,model.sex);
+      NSLog(@"ID : %ld, name : %@, age : %@ ",(long)model.id,model.name,model.sex);
         
     }
 }
